feat: created welcome page and added the routes
