+++ src/routes/admin/manage-reservations/+page.svelte
@@ -1,0 +1,13 @@
+<script>
+</script>
+
+<main>
+	<p>🚧</p>
+</main>
+
+<style>
+	p {
+		font-size: 8rem;
+		text-align: center;
+	}
+</style>
